+++ src/App.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useMemo } from "react";
 import { themes, type ThemeName } from "./styles/theme";
 import { Header } from "./components/Header";
 import { ErrorMessage } from "./components/ErrorMessage";
+import { MovieForm } from "./components/MovieForm";
 
 interface Movie {
   id: number;
@@ -120,105 +121,22 @@ function App() {
         currentTheme={currentTheme}
       />
 
-      <div
-        style={{
-          marginBottom: "24px",
-          padding: "20px",
-          borderRadius: "12px",
-          backgroundColor: currentTheme.componentBg,
-          border: `1px solid ${currentTheme.border}`,
-        }}
-      >
-        <h2>영화 추가</h2>
-        <form
-          onSubmit={handleAddMovie}
-          style={{ display: "flex", gap: "12px", flexWrap: "wrap" }}
-        >
-          <input
-            type="text"
-            placeholder="제목"
-            value={newTitle}
-            onChange={(e) => setNewTitle(e.target.value)}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-            }}
-          />
-          <input
-            type="text"
-            placeholder="감독"
-            value={newDirector}
-            onChange={(e) => setNewDirector(e.target.value)}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-            }}
-          />
-          <input
-            type="number"
-            placeholder="연도"
-            value={newYear}
-            onChange={(e) => setNewYear(Number(e.target.value))}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-              width: "80px",
-            }}
-          />
-          <input
-            type="text"
-            placeholder="장르"
-            value={newGenre}
-            onChange={(e) => setNewGenre(e.target.value)}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-            }}
-          />
-          <input
-            type="number"
-            placeholder="평점"
-            value={newRating}
-            onChange={(e) => {
-              const val = Number(e.target.value);
-              if (val >= 0 && val <= 10) setNewRating(val);
-            }}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-              width: "100px",
-            }}
-          />
-          <button
-            type="submit"
-            style={{
-              padding: "8px 16px",
-              backgroundColor: currentTheme.buttonBg,
-              color: currentTheme.buttonText,
-              border: "none",
-              borderRadius: "8px",
-              cursor: "pointer",
-            }}
-          >
-            추가
-          </button>
-        </form>
-      </div>
+      {/* MovieForm 호출 */}
+      <MovieForm
+        newTitle={newTitle}
+        setNewTitle={setNewTitle}
+        newDirector={newDirector}
+        setNewDirector={setNewDirector}
+        newYear={newYear}
+        setNewYear={setNewYear}
+        newGenre={newGenre}
+        setNewGenre={setNewGenre}
+        newRating={newRating}
+        setNewRating={setNewRating}
+        onSubmit={handleAddMovie}
+        currentTheme={currentTheme}
+      />
+
       <div
         style={{
           padding: "20px",
